Add context menu item
This implementation is very incomplete, as I found no way to acquire
the link the user clicked on (if it was a link afterall) in a
reliable way.

ATM, i'm using jQuery and search through the document for a link
containing getSelection(); but that might yield multiple results, or
false positives.